feat:  separate in new file && no log if no themes
